Effect

Description

Types and functions wrapping Fixpoint.Internal into an user-facing API

Documentation

type Effect = (Type -> Type) -> Type -> Type

class EffectInject (op :: Effect) (ops :: [Effect]) | ops -> op

Is the given type of effects an element of the list of effect types? Users won't write instances of this type, they're generated automatically.

In writing "reification" instances, a useful idiom is something like

instance EffectInject (WriterEffect w) ops => MonadWriter (AST ops) where
  ...

which will define "writer-like syntax" for all lists of effects that contain 'WriterEffect w'. This idiom is also the reason for the functional dependency in the class: ops must contain enough information to determine op.

Minimal complete definition

effectInject

type AST ops = Fixpoint Freer (JoinedEffects ops)

An abstract syntax tree with effects of type op

astInject :: EffectInject op ops => op (AST ops) a -> AST ops a

convenience function to build the "singleton" AST corresponding to one effect

class InterpretEffect m op where

Users will write an instance of this class if they want to define the "standard" interpretation of their effect type.

Methods

interpretEffect :: (forall b. AST ops b -> m b) -> op (AST ops) a -> m a

Given a function that describes how to interpret ASTs and an effect that may contain zero or more ASTs, return the interpretation od the effect.

class InterpretEffects m ops

The constraint that all effects ops are interpretable (in the sense of InterpretEffect) in m

Minimal complete definition

interpretEffects

interpretAST :: (Monad m, InterpretEffects m ops) => AST ops a -> m a

If all effect types in the AST have an 'InterpretEffect m' instance, this function will interpret the AST.

note: This function is morally the same as fromFixpoint, it's just that the instance resolution machinery (especially around the type-level lists of effect states) obfuscates it.

class InterpretEffectStateful (t :: Type -> Type) m op where

Users will write an instance of this class if they want to interpret effects of type op in m, while also passing a state of type x form step to step. This state will control how effects are implemented.

For example in Fixpoint.Examples, I play around with the type of LTL formulas for x.

Methods

interpretEffectStateful :: (forall b y. t y -> AST ops b -> m (b, t y)) -> t x -> op (AST ops) a -> m (a, t x)

Given a function that describes how to interpret ASTs statefully, a current "interpretation state", and an effect, return the interpretation of the effect.

class InterpretEffectsStateful t m ops where

The constraint that all effects ops are statefully interpretable (in the sense of InterpretEffectStateful) in m.

Methods

interpretEffectsStateful :: ConstraintList (InterpretEffectStateful t m) ops

interpretASTStateful :: (Monad m, InterpretEffectsStateful t m ops) => t x -> AST ops a -> m (a, t x)

If all effect types in the AST have an InterpretEffectStateful instance, this function will interpret the AST, starting at a given initial "interpretation state"

note: This function is morally the same as fromFixpointState, it's just that the instance resolution machinery (especially around the type-level lists of effect states) obfuscates it.

Internal

data JoinedEffects (ops :: [Effect]) :: Effect where

Constructors

JoinedEffectsHere :: op m a -> JoinedEffects (op ': ops) m a

JoinedEffectsThere :: JoinedEffects ops m a -> JoinedEffects (op ': ops) m a